refactor: ArgumentResolver에 애노테이션 사용
