+++ src/main/java/nextstep/member/config/WebMvcConfig.java
@@ -2,6 +2,7 @@ package nextstep.member.config;
 
 import lombok.RequiredArgsConstructor;
 import nextstep.member.application.JwtTokenProvider;
+import nextstep.member.config.argument.resolver.FindMemberOfMineRequestArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
