+++ src/main/java/nextstep/member/config/argument/annotation/FindMemberOfMine.java
@@ -1,0 +1,11 @@
+package nextstep.member.config.argument.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.PARAMETER)
+public @interface FindMemberOfMine {
+}
