+++ src/main/java/nextstep/member/config/argument/resolver/FindMemberOfMineRequestArgumentResolver.java
@@ -1,8 +1,9 @@
-package nextstep.member.config;
+package nextstep.member.config.argument.resolver;
 
 import lombok.RequiredArgsConstructor;
 import nextstep.member.application.JwtTokenProvider;
 import nextstep.member.application.dto.FindMemberOfMineRequest;
+import nextstep.member.config.argument.annotation.FindMemberOfMine;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
@@ -15,7 +16,7 @@ public class FindMemberOfMineRequestArgumentResolver implements HandlerMethodArg
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.getParameterType() == FindMemberOfMineRequest.class;
+        return parameter.getParameterAnnotation(FindMemberOfMine.class) != null;
     }
 
     @Override
@@ -23,7 +24,6 @@ public class FindMemberOfMineRequestArgumentResolver implements HandlerMethodArg
                                   ModelAndViewContainer mavContainer,
                                   NativeWebRequest webRequest,
                                   WebDataBinderFactory binderFactory) throws Exception {
-
         String token = webRequest.getHeader("authorization");
         return new FindMemberOfMineRequest(jwtTokenProvider.getPrincipal(token), jwtTokenProvider.getRoles(token));
     }
